Move loggin Clas to it's own file
