+++ src/KerbalKonstructs.csproj
@@ -30,6 +30,9 @@
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
+  <PropertyGroup>
+    <RunPostBuildEvent>OnBuildSuccess</RunPostBuildEvent>
+  </PropertyGroup>
   <ItemGroup>
     <Reference Include="Assembly-CSharp, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
@@ -84,6 +87,7 @@
     <Compile Include="SpaceCenters\SpaceCenterManager.cs" />
     <Compile Include="StaticObjects\StaticUtils.cs" />
     <Compile Include="UI\UIMain.cs" />
+    <Compile Include="Utilities\Log.cs" />
     <Compile Include="Wrappers\StageRecoveryWrapper.cs" />
     <Compile Include="Nations\KerbinNation.cs" />
     <Compile Include="StaticObjects\StaticModel.cs" />
@@ -120,8 +124,7 @@
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
-    <PostBuildEvent>
-    </PostBuildEvent>
+    <PostBuildEvent>rmdir /q /s  $(ProjectDir)\obj</PostBuildEvent>
   </PropertyGroup>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
